requirements.txt
numpy
scikit-learn
joblib
openml
streamlit
pillow
opencv-python
streamlit-drawable-canvas
